requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
python-dotenv
urllib3<2